moved all NRPE related functions to nrpe_helpers.py
