+++ tests/test_openstack_integrator.py
@@ -31,26 +31,26 @@ def test_initial_nrpe_config_failed(mock_update_nrpe_config):
     mock_update_nrpe_config.assert_called_once()
 
 
-@mock.patch("charms.layer.openstack.write_nagios_openstack_cnf")
-def test_update_nrpe_config(mock_write_nagios_openstack_cnf, openstack_config):
+@mock.patch("nrpe_helpers.write_nagios_openstack_cnf")
+def test_update_nrpe_config(mock_write_nagios_openstack_cnf):
     nrpe.NRPE.return_value = mock_nrpe_setup = MagicMock()
     NAGIOS_PLUGINS_DIR.__fspath__.return_value = "/test/"
-    config.return_value = openstack_config
+    config.return_value = mock_config = MagicMock()
 
     # no NRPE checks were added / deleted
-    openstack_config.changed.return_value = False
-    openstack_config.get.side_effect = {}.get
+    mock_config.changed.return_value = False
+    mock_config.get.side_effect = {}.get
     openstack.update_nrpe_config()
     mock_nrpe_setup.add_check.assert_not_called()
     mock_nrpe_setup.remove_check.assert_not_called()
     mock_write_nagios_openstack_cnf.assert_called_once_with()
     mock_write_nagios_openstack_cnf.reset_mock()
     mock_nrpe_setup.reset_mock()
-    openstack_config.reset_mock()
+    mock_config.reset_mock()
 
     # add NRPE check for kubernetes subnet
-    openstack_config.changed.return_value = True
-    openstack_config.get.side_effect = {"subnet-id": "1234"}.get
+    mock_config.changed.return_value = True
+    mock_config.get.side_effect = {"subnet-id": "1234"}.get
     openstack.update_nrpe_config()
     mock_nrpe_setup.add_check.assert_called_once_with(
         shortname="kubernetes_subnet",
@@ -60,11 +60,11 @@ def test_update_nrpe_config(mock_write_nagios_openstack_cnf, openstack_config):
     )
     assert mock_nrpe_setup.remove_check.call_count == 7
     mock_nrpe_setup.reset_mock()
-    openstack_config.reset_mock()
+    mock_config.reset_mock()
 
     # add NRPE check for servers
-    openstack_config.changed.return_value = True
-    openstack_config.get.side_effect = {"nrpe-server-ids": "1,2,3"}.get
+    mock_config.changed.return_value = True
+    mock_config.get.side_effect = {"nrpe-server-ids": "1,2,3"}.get
     openstack.update_nrpe_config()
     mock_nrpe_setup.add_check.assert_called_once_with(
         shortname="openstack_servers",
@@ -74,12 +74,12 @@ def test_update_nrpe_config(mock_write_nagios_openstack_cnf, openstack_config):
     )
     assert mock_nrpe_setup.remove_check.call_count == 7
     mock_nrpe_setup.reset_mock()
-    openstack_config.reset_mock()
+    mock_config.reset_mock()
 
     # add NRPE check for servers with skip-ids
-    openstack_config.changed.return_value = True
-    openstack_config.get.side_effect = {"nrpe-server-ids": "all",
-                                        "nrpe-skip-server-ids": "1,2"}.get
+    mock_config.changed.return_value = True
+    mock_config.get.side_effect = {
+        "nrpe-server-ids": "all", "nrpe-skip-server-ids": "1,2"}.get
     openstack.update_nrpe_config()
     mock_nrpe_setup.add_check.assert_called_once_with(
         shortname="openstack_servers",
@@ -89,32 +89,31 @@ def test_update_nrpe_config(mock_write_nagios_openstack_cnf, openstack_config):
     )
     assert mock_nrpe_setup.remove_check.call_count == 7
     mock_nrpe_setup.reset_mock()
-    openstack_config.reset_mock()
+    mock_config.reset_mock()
 
     # raise a ValueError while adding NRPE check
-    openstack_config.changed.return_value = True
-    openstack_config.get.side_effect = {"nrpe-server-ids": "1,2",
-                                        "nrpe-skip-server-ids": "1"}.get
+    mock_config.changed.return_value = True
+    mock_config.get.side_effect = {
+        "nrpe-server-ids": "1,2", "nrpe-skip-server-ids": "1"}.get
     with pytest.raises(ValueError):
         openstack.update_nrpe_config()
 
-    openstack_config.get.side_effect = {"subnet-id": "all"}.get
+    mock_config.get.side_effect = {"subnet-id": "all"}.get
 
     with pytest.raises(ValueError):
         openstack.update_nrpe_config()
 
-    openstack_config.reset_mock()
+    mock_config.reset_mock()
 
 
-@mock.patch("charms.layer.openstack.remove_nagios_openstack_cnf")
-def test_remove_nrpe_config(mock_remove_nagios_openstack_cnf,
-                            openstack_config):
+@mock.patch("nrpe_helpers.remove_nagios_openstack_cnf")
+def test_remove_nrpe_config(mock_remove_nagios_openstack_cnf):
     nrpe.NRPE.return_value = mock_nrpe_setup = MagicMock()
     NAGIOS_PLUGINS_DIR.__fspath__.return_value = "/test/"
-    config.return_value = openstack_config
+    config.return_value = mock_config = MagicMock()
 
     # no NRPE checks were removed
-    openstack_config.get.side_effect = {}.get
+    mock_config.get.side_effect = {}.get
     openstack.remove_nrpe_config()
     mock_nrpe_setup.remove_check.assert_not_called()
     mock_remove_nagios_openstack_cnf.assert_called_once_with()
@@ -122,7 +121,7 @@ def test_remove_nrpe_config(mock_remove_nagios_openstack_cnf,
     mock_nrpe_setup.reset_mock()
 
     # remove NRPE check for kubernetes network
-    openstack_config.get.side_effect = {"nrpe-server-ids": "1,2,3"}.get
+    mock_config.get.side_effect = {"nrpe-server-ids": "1,2,3"}.get
     openstack.remove_nrpe_config()
     mock_nrpe_setup.remove_check.assert_called_once_with(
         shortname="openstack_servers", description="",
